+++ hector_gamepad_manager/include/hector_gamepad_manager/hector_gamepad_manager.hpp
@@ -1,7 +1,7 @@
 #ifndef HECTOR_GAMEPAD_MANAGER_HECTOR_GAMEPAD_MANAGER_HPP
 #define HECTOR_GAMEPAD_MANAGER_HECTOR_GAMEPAD_MANAGER_HPP
 
-#include "hector_gamepad_manager_interface/gamepad_plugin_interface.hpp"
+#include "hector_gamepad_plugin_interface/gamepad_plugin_interface.hpp"
 
 #include <pluginlib/class_loader.hpp>
 #include <rclcpp/rclcpp.hpp>
@@ -14,7 +14,7 @@ namespace hector_gamepad_manager
 class HectorGamepadManager
 {
 public:
-  using GamepadFunctionPlugin = hector_gamepad_manager_interface::GamepadFunctionPlugin;
+  using GamepadFunctionPlugin = hector_gamepad_plugin_interface::GamepadFunctionPlugin;
   // Constructor
   explicit HectorGamepadManager( const rclcpp::Node::SharedPtr &node );
