+++ hector_gamepad_manager/package.xml
@@ -9,7 +9,7 @@
 
   <buildtool_depend>ament_cmake</buildtool_depend>
 
-  <depend>hector_gamepad_manager_interface</depend>
+  <depend>hector_gamepad_plugin_interface</depend>
   <depend>pluginlib</depend>
   <depend>rclcpp</depend>
   <depend>sensor_msgs</depend>
