+++ hector_gamepad_manager_plugins/CMakeLists.txt
@@ -18,7 +18,7 @@ find_package(ament_cmake REQUIRED)
 set(DEPENDENCIES
         controller_manager_msgs
         geometry_msgs
-        hector_gamepad_manager_interface
+        hector_gamepad_plugin_interface
         hector_ros2_utils
         moveit_msgs
         pluginlib
