+++ hector_gamepad_manager_plugins/include/hector_gamepad_manager_plugins/manipulation_plugin.hpp
@@ -5,15 +5,15 @@
 #define HECTOR_GAMEPAD_MANAGER_PLUGINS_MANIPULATION_PLUGIN_HPP
 
 #include <geometry_msgs/msg/twist_stamped.hpp>
-#include <hector_gamepad_manager_interface/gamepad_plugin_interface.hpp>
 #include <hector_gamepad_manager_plugins/controller_helper.hpp>
+#include <hector_gamepad_plugin_interface/gamepad_plugin_interface.hpp>
 #include <hector_ros2_utils/parameters/reconfigurable_parameter.hpp>
 #include <std_msgs/msg/float64.hpp>
 #include <std_srvs/srv/set_bool.hpp>
 
 namespace hector_gamepad_manager_plugins
 {
-class ManipulationPlugin final : public hector_gamepad_manager_interface::GamepadFunctionPlugin
+class ManipulationPlugin final : public hector_gamepad_plugin_interface::GamepadFunctionPlugin
 {
 public:
   void initialize( const rclcpp::Node::SharedPtr &node ) override;
