+++ hector_gamepad_manager_plugins/package.xml
@@ -11,7 +11,7 @@
 
   <depend>controller_manager_msgs</depend>
   <depend>geometry_msgs</depend>
-  <depend>hector_gamepad_manager_interface</depend>
+  <depend>hector_gamepad_plugin_interface</depend>
   <depend>hector_ros2_utils</depend>
   <depend>moveit_msgs</depend>
   <depend>pluginlib</depend>
