+++ hector_gamepad_manager_plugins/plugins.xml
@@ -1,15 +1,15 @@
 <?xml version="1.0"?>
 <library path="hector_gamepad_manager_plugins">
-    <class type="hector_gamepad_manager_plugins::DrivePlugin" base_class_type="hector_gamepad_manager_interface::GamepadFunctionPlugin">
+    <class type="hector_gamepad_manager_plugins::DrivePlugin" base_class_type="hector_gamepad_plugin_interface::GamepadFunctionPlugin">
         <description>Drive functionality plugin</description>
     </class>
-    <class type="hector_gamepad_manager_plugins::ManipulationPlugin" base_class_type="hector_gamepad_manager_interface::GamepadFunctionPlugin">
+    <class type="hector_gamepad_manager_plugins::ManipulationPlugin" base_class_type="hector_gamepad_plugin_interface::GamepadFunctionPlugin">
         <description>Manipulation functionality plugin</description>
     </class>
-    <class type="hector_gamepad_manager_plugins::FlipperPlugin" base_class_type="hector_gamepad_manager_interface::GamepadFunctionPlugin">
+    <class type="hector_gamepad_manager_plugins::FlipperPlugin" base_class_type="hector_gamepad_plugin_interface::GamepadFunctionPlugin">
         <description>Flipper functionality plugin</description>
     </class>
-    <class type="hector_gamepad_manager_plugins::MoveitPlugin" base_class_type="hector_gamepad_manager_interface::GamepadFunctionPlugin">
+    <class type="hector_gamepad_manager_plugins::MoveitPlugin" base_class_type="hector_gamepad_plugin_interface::GamepadFunctionPlugin">
         <description>Interface to moveit planners.</description>
     </class>
 </library>
