+++ hector_gamepad_manager_plugins/src/moveit_plugin.cpp
@@ -259,4 +259,4 @@ std::pair<std::string, std::string> MoveitPlugin::fromGroupPoseName( const std::
 #include <pluginlib/class_list_macros.hpp>
 
 PLUGINLIB_EXPORT_CLASS( hector_gamepad_manager_plugins::MoveitPlugin,
-                        hector_gamepad_manager_interface::GamepadFunctionPlugin )
+                        hector_gamepad_plugin_interface::GamepadFunctionPlugin )
