+++ hector_gamepad_plugin_interface/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>hector_gamepad_manager_interface</name>
+  <name>hector_gamepad_plugin_interface</name>
   <version>0.0.0</version>
   <description>Provides plugin base class for gamepad manager plugins</description>
