+++ hector_gamepad_plugin_interface/src/gamepad_plugin_interface.cpp
@@ -1,0 +1,10 @@
+//
+// Created by aljoscha-schmidt on 3/27/25.
+//
+
+#include <hector_gamepad_plugin_interface/gamepad_plugin_interface.hpp>
+
+namespace hector_gamepad_plugin_interface
+{
+
+} // namespace hector_gamepad_plugin_interface
